LayoutTests/imported/w3c/web-platform-tests/svg/painting/parsing/text-rendering-computed-expected.txt


PASS Property text-rendering value 'auto' computes to 'auto' 
FAIL Property text-rendering value 'optimizespeed' computes to 'optimizespeed' assert_equals: expected "optimizespeed" but got "optimizeSpeed"
FAIL Property text-rendering value 'optimizelegibility' computes to 'optimizelegibility' assert_equals: expected "optimizelegibility" but got "optimizeLegibility"
FAIL Property text-rendering value 'geometricprecision' computes to 'geometricprecision' assert_equals: expected "geometricprecision" but got "geometricPrecision"
